requirements.txt
absl-py==2.1.0
alembic==1.13.1
cffi==1.16.0
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
docstring_parser==0.16
fasteners==0.19
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
glfw==2.7.0
greenlet==3.0.3
grpcio==1.64.1
imageio==2.34.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
mujoco-py==2.1.2.14
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.0.0
shtab==1.7.1
six==1.16.0
SQLAlchemy==2.0.30
sympy==1.12.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.3.1
torch-summary==1.4.5
torch-tb-profiler==0.4.3
tqdm==4.66.4
typing_extensions==4.12.2
tyro==0.8.4
tzdata==2024.1
Werkzeug==3.0.3
